rename api router to weather
