+++ src/routes/index.mjs
@@ -1,11 +1,11 @@
 import { Router } from "express";
-import apiRouter from "./api.mjs";
+import weatherRouter from "./weather.mjs";
 import infoRouter from "./info.mjs"
 import authRouter from "./auth.mjs";
 
 const router = Router();
 
-router.use(apiRouter);
+router.use(weatherRouter);
 router.use(infoRouter);
 router.use(authRouter);
